+++ scripts/gripper_action_client.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# Copyright (c) 2013, Rethink Robotics
+# Copyright (c) 2013-2014, Rethink Robotics
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -43,6 +43,8 @@ from control_msgs.msg import (
 )
 
 import baxter_interface
+
+from baxter_interface import CHECK_VERSION
 
 
 class GripperClient(object):
@@ -103,7 +105,7 @@ def main():
     print("Initializing node... ")
     rospy.init_node("rsdk_gripper_action_client_%s" % (gripper,))
     print("Getting robot state... ")
-    rs = baxter_interface.RobotEnable()
+    rs = baxter_interface.RobotEnable(CHECK_VERSION)
     print("Enabling robot... ")
     rs.enable()
     print("Running. Ctrl-c to quit")