+++ scripts/joint_position_joystick.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# Copyright (c) 2013, Rethink Robotics
+# Copyright (c) 2013-2014, Rethink Robotics
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -36,6 +36,8 @@ import rospy
 
 import baxter_interface
 import baxter_external_devices
+
+from baxter_interface import CHECK_VERSION
 
 
 def rotate(l):
@@ -74,8 +76,8 @@ def map_joystick(joystick):
     """
     left = baxter_interface.Limb('left')
     right = baxter_interface.Limb('right')
-    grip_left = baxter_interface.Gripper('left')
-    grip_right = baxter_interface.Gripper('right')
+    grip_left = baxter_interface.Gripper('left', CHECK_VERSION)
+    grip_right = baxter_interface.Gripper('right', CHECK_VERSION)
     lcmd = {}
     rcmd = {}
 
@@ -204,7 +206,7 @@ key bindings.
     print("Initializing node... ")
     rospy.init_node("rsdk_joint_position_joystick")
     print("Getting robot state... ")
-    rs = baxter_interface.RobotEnable()
+    rs = baxter_interface.RobotEnable(CHECK_VERSION)
     init_state = rs.state().enabled
 
     def clean_shutdown():