+++ scripts/joint_recorder.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# Copyright (c) 2013, Rethink Robotics
+# Copyright (c) 2013-2014, Rethink Robotics
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -34,6 +34,8 @@ import rospy
 import baxter_interface
 from baxter_examples import JointRecorder
 
+from baxter_interface import CHECK_VERSION
+
 
 def main():
     """RSDK Joint Recorder Example
@@ -65,15 +67,15 @@ Related examples:
         help='the file name to record to'
     )
     parser.add_argument(
-        '-r', '--record-rate', type=int, default=10, metavar='RECORDRATE',
-        help='rate at which to record'
+        '-r', '--record-rate', type=int, default=100, metavar='RECORDRATE',
+        help='rate at which to record (default: 100)'
     )
     args = parser.parse_args(rospy.myargv()[1:])
 
     print("Initializing node... ")
     rospy.init_node("rsdk_joint_recorder")
     print("Getting robot state... ")
-    rs = baxter_interface.RobotEnable()
+    rs = baxter_interface.RobotEnable(CHECK_VERSION)
     print("Enabling robot... ")
     rs.enable()