+++ CHANGELOG.rst
@@ -1,3 +1,14 @@
+1.0.0 (2014-5-1)
+---------------------------------
+- Adds joint_position_waypoints example program
+- Updates ik_service_client to validate unpacked results and seed types
+- Updates gripper_cuff_control to automatically calibrate on gripper type change
+- Updates navigator_io to use navigator wheel_changed signal
+- Updates all examples to verify robot software version by default when enabling
+- Updates all examples using gripper to verify gripper firmware version
+- Updates joint_recorder to record at 100Hz
+- Updates joint_velocity_wobbler to run at 500Hz
+
 0.7.0 (2013-11-21)
 ---------------------------------
 - Creation of baxter_examples repository from sdk-examples/examples.